requirements.txt
Flask==3.0.3
flask-oidc==1.4.0
requests==2.32.3
python-dotenv==1.0.1
